+++ core/ledger/kvledger/txmgmt/privacyenabledstate/types.go
@@ -9,38 +9,9 @@ package privacyenabledstate
 import (
 	"fmt"
 
-	"github.com/hyperledger/fabric/core/ledger/cceventmgmt"
 	"github.com/hyperledger/fabric/core/ledger/internal/version"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/statedb"
 )
-
-// DBProvider provides handle to a PvtVersionedDB
-type DBProvider interface {
-	// GetDBHandle returns a handle to a PvtVersionedDB
-	GetDBHandle(id string) (DB, error)
-	// Close closes all the PvtVersionedDB instances and releases any resources held by VersionedDBProvider
-	Close()
-}
-
-// DB extends VersionedDB interface. This interface provides additional functions for managing private data state
-type DB interface {
-	statedb.VersionedDB
-	IsBulkOptimizable() bool
-	LoadCommittedVersionsOfPubAndHashedKeys(pubKeys []*statedb.CompositeKey, hashedKeys []*HashedCompositeKey) error
-	GetCachedKeyHashVersion(namespace, collection string, keyHash []byte) (*version.Height, bool)
-	ClearCachedVersions()
-	GetChaincodeEventListener() cceventmgmt.ChaincodeLifecycleEventListener
-	GetPrivateData(namespace, collection, key string) (*statedb.VersionedValue, error)
-	GetPrivateDataHash(namespace, collection, key string) (*statedb.VersionedValue, error)
-	GetValueHash(namespace, collection string, keyHash []byte) (*statedb.VersionedValue, error)
-	GetKeyHashVersion(namespace, collection string, keyHash []byte) (*version.Height, error)
-	GetPrivateDataMultipleKeys(namespace, collection string, keys []string) ([]*statedb.VersionedValue, error)
-	GetPrivateDataRangeScanIterator(namespace, collection, startKey, endKey string) (statedb.ResultsIterator, error)
-	GetStateMetadata(namespace, key string) ([]byte, error)
-	GetPrivateDataMetadataByHash(namespace, collection string, keyHash []byte) ([]byte, error)
-	ExecuteQueryOnPrivateData(namespace, collection, query string) (statedb.ResultsIterator, error)
-	ApplyPrivacyAwareUpdates(updates *UpdateBatch, height *version.Height) error
-}
 
 // PvtdataCompositeKey encloses Namespace, CollectionName and Key components
 type PvtdataCompositeKey struct {
